added two provenance columns by AstTransform + modifications to RamTranslator and ReadStream
