+++ src/RamExecutor.cpp
@@ -195,6 +195,9 @@ RamDomain eval(const RamValue& value, RamEnvironment& env, const EvalContext& ct
                 case BinaryOp::LOR: {
                     return visit(op.getLHS()) || visit(op.getRHS());
                 }
+                case BinaryOp::MAX: {
+                    return std::max(visit(op.getLHS()), visit(op.getRHS()));
+                }
 
                 // strings
                 case BinaryOp::CAT: {
@@ -1862,6 +1865,11 @@ public:
             }
             case BinaryOp::LOR: {
                 out << "(" << print(op.getLHS()) << ") || (" << print(op.getRHS()) << ")";
+                break;
+            }
+            case BinaryOp::MAX: {
+                out << "(AstDomain)(std::max((AstDomain)" << print(op.getLHS()) << ","
+                    << "(AstDomain)" << print(op.getRHS()) << "))";
                 break;
             }
