+++ src/ReadStreamCSV.h
@@ -103,6 +103,14 @@ public:
                 }
             }
         }
+
+        // add two provenance columns
+        if (Global::config().has("provenance")) {
+            tuple[symbolMask.getArity() - 2] = 0;
+            tuple[symbolMask.getArity() - 1] = 0;
+            columnsFilled += 2;
+        }
+
         if (columnsFilled != symbolMask.getArity()) {
             std::stringstream errorMessage;
             errorMessage << "Values missing in line " << lineNumber << "; ";
